pyproject.toml
[tool.poetry]
name = "savage-ai-im"
version = "0.1.0"
description = ""
authors = ["freyamade"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
jupyter = "^1.1.1"
langchain-core = "^0.3.15"
langchain-community = "^0.3.5"
langgraph = "^0.2.45"
tavily-python = "^0.5.0"
langgraph-checkpoint-sqlite = "^2.0.1"
langchain-google-vertexai = "^2.0.7"
discord-py = "^2.4.0"
sentry-sdk = "^2.18.0"
structlog = "^24.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
